batch_jobs_to_bigquery/scrape_triggered_gcs_to_bq/TriggerFunction/requirements.txt

google-cloud-storage
google-api-python-client